feat: package restructure and adding in geocoder.
