+++ app/src/main/java/com/example/recipe_app_compose/core/navigation/RecipeApp.kt
@@ -8,16 +8,16 @@ import androidx.lifecycle.viewmodel.compose.viewModel
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
-import com.example.recipe_app_compose.domain.model.category.Category
-import com.example.recipe_app_compose.presentation.view.AccountScreen
-import com.example.recipe_app_compose.presentation.view.DetailScreen
-import com.example.recipe_app_compose.presentation.view.FavoritesScreen
-import com.example.recipe_app_compose.presentation.view.InfoScreen
-import com.example.recipe_app_compose.presentation.view.IngredientScreen
-import com.example.recipe_app_compose.presentation.view.RandomMealPage
-import com.example.recipe_app_compose.presentation.view.RecipeScreen
-import com.example.recipe_app_compose.presentation.view.SettingsScreen
-import com.example.recipe_app_compose.presentation.viewmodel.RecipeViewModel
+import com.example.recipe_app_compose.features.categories.domain.model.category.Category
+import com.example.recipe_app_compose.features.categories.presentation.view.AccountScreen
+import com.example.recipe_app_compose.features.categories.presentation.view.DetailScreen
+import com.example.recipe_app_compose.features.categories.presentation.view.FavoritesScreen
+import com.example.recipe_app_compose.features.categories.presentation.view.InfoScreen
+import com.example.recipe_app_compose.features.categories.presentation.view.IngredientScreen
+import com.example.recipe_app_compose.features.categories.presentation.view.RandomMealPage
+import com.example.recipe_app_compose.features.categories.presentation.view.RecipeScreen
+import com.example.recipe_app_compose.features.categories.presentation.view.SettingsScreen
+import com.example.recipe_app_compose.features.categories.presentation.viewmodel.RecipeViewModel
 
 /*
 File for Navigation.
@@ -52,8 +52,8 @@ fun RecipeApp(navController: NavHostController, modifier: Modifier) {
         ) {
             val category =
                 navController
-                    .previousBackStackEntry?.savedStateHandle?.get<Category>("nav")
-                    ?: Category(
+                    .previousBackStackEntry?.savedStateHandle?.get<com.example.recipe_app_compose.features.categories.domain.model.category.Category>("nav")
+                    ?: com.example.recipe_app_compose.features.categories.domain.model.category.Category(
                         "",
                         "",
                         "",
