--- app/src/main/java/com/example/recipe_app_compose/data/api/RetrofitInstance.kt
@@ -1,35 +1,0 @@
-package com.example.recipe_app_compose.data.api
-
-import com.example.recipe_app_compose.core.util.Constants.BASE_URL
-import okhttp3.OkHttpClient
-import okhttp3.logging.HttpLoggingInterceptor
-import retrofit2.Retrofit
-import retrofit2.converter.gson.GsonConverterFactory
-import java.util.concurrent.TimeUnit
-
-object RetrofitInstance {
-    private fun provideHttpInterceptor(): OkHttpClient {
-        val interceptor = HttpLoggingInterceptor().apply {
-            this.level = HttpLoggingInterceptor.Level.BODY
-        }
-
-        val client = OkHttpClient.Builder().apply {
-            this.addInterceptor(interceptor)
-                .connectTimeout(10, TimeUnit.SECONDS)
-                .readTimeout(10, TimeUnit.SECONDS)
-                .writeTimeout(10, TimeUnit.SECONDS)
-        }.build()
-        return client
-    }
-
-    private val gson: GsonConverterFactory = GsonConverterFactory.create()
-    private val retrofit: Retrofit = Retrofit.Builder()
-        .baseUrl(BASE_URL)
-        .addConverterFactory(gson)
-        .client(provideHttpInterceptor())
-        .build()
-
-    val apiService: ApiService by lazy {
-        retrofit.create(ApiService::class.java)
-    }
-}
