--- app/src/main/java/com/example/recipe_app_compose/domain/model/category/Category.kt
@@ -1,21 +1,0 @@
-package com.example.recipe_app_compose.domain.model.category
-
-import android.os.Parcelable
-import kotlinx.parcelize.Parcelize
-
-@Parcelize
-data class Category(
-    val idCategory: String,
-    val strCategory: String,
-    val strCategoryThumb: String,
-    val strCategoryDescription: String
-) : Parcelable
-
-/*
-{
-      "idCategory": "1",
-      "strCategory": "Beef",
-      "strCategoryThumb": "https://www.themealdb.com/images/category/beef.png",
-      "strCategoryDescription": "Beef is the culinary name for meat from cattle, particularly skeletal muscle. Humans have been eating beef since prehistoric times.[1] Beef is a source of high-quality protein and essential nutrients.[2]"
-    },
- */
