--- app/src/main/java/com/example/recipe_app_compose/domain/model/category/CategoryResponse.kt
@@ -1,9 +1,0 @@
-package com.example.recipe_app_compose.domain.model.category
-
-import android.os.Parcelable
-import kotlinx.parcelize.Parcelize
-
-@Parcelize
-data class CategoryResponse(
-    val categories: List<Category>
-) : Parcelable
