--- app/src/main/java/com/example/recipe_app_compose/domain/model/categorymeal/CategoryMealResponse.kt
@@ -1,9 +1,0 @@
-package com.example.recipe_app_compose.domain.model.categorymeal
-
-import android.os.Parcelable
-import kotlinx.parcelize.Parcelize
-
-@Parcelize
-data class CategoryMealResponse(
-    val meals: List<CategoryMeal>
-) : Parcelable
