--- app/src/main/java/com/example/recipe_app_compose/domain/model/ingredient/IngredientResponse.kt
@@ -1,9 +1,0 @@
-package com.example.recipe_app_compose.domain.model.ingredient
-
-import android.os.Parcelable
-import kotlinx.parcelize.Parcelize
-
-@Parcelize
-data class IngredientResponse(
-    val meals: List<Ingredient>
-) : Parcelable
