--- app/src/main/java/com/example/recipe_app_compose/domain/model/location/LocationData.kt
@@ -1,6 +1,0 @@
-package com.example.recipe_app_compose.domain.model.location
-
-data class LocationData(
-    val latitude: Double,
-    val longitude: Double
-)
