--- app/src/main/java/com/example/recipe_app_compose/domain/model/randommeal/RandomMealResponse.kt
@@ -1,9 +1,0 @@
-package com.example.recipe_app_compose.domain.model.randommeal
-
-import android.os.Parcelable
-import kotlinx.parcelize.Parcelize
-
-@Parcelize
-data class RandomMealResponse(
-    val meals: List<RandomMeal>
-) : Parcelable
