+++ app/src/main/java/com/example/recipe_app_compose/features/categories/domain/model/categorymeal/CategoryMealResponse.kt
@@ -1,0 +1,9 @@
+package com.example.recipe_app_compose.features.categories.domain.model.categorymeal
+
+import android.os.Parcelable
+import kotlinx.parcelize.Parcelize
+
+@Parcelize
+data class CategoryMealResponse(
+    val meals: List<com.example.recipe_app_compose.features.categories.domain.model.categorymeal.CategoryMeal>
+) : Parcelable
