+++ app/src/main/java/com/example/recipe_app_compose/features/categories/domain/model/ingredient/IngredientResponse.kt
@@ -1,0 +1,9 @@
+package com.example.recipe_app_compose.features.categories.domain.model.ingredient
+
+import android.os.Parcelable
+import kotlinx.parcelize.Parcelize
+
+@Parcelize
+data class IngredientResponse(
+    val meals: List<com.example.recipe_app_compose.features.categories.domain.model.ingredient.Ingredient>
+) : Parcelable
