+++ app/src/main/java/com/example/recipe_app_compose/features/categories/domain/model/randommeal/RandomMealResponse.kt
@@ -1,0 +1,9 @@
+package com.example.recipe_app_compose.features.categories.domain.model.randommeal
+
+import android.os.Parcelable
+import kotlinx.parcelize.Parcelize
+
+@Parcelize
+data class RandomMealResponse(
+    val meals: List<com.example.recipe_app_compose.features.categories.domain.model.randommeal.RandomMeal>
+) : Parcelable
