+++ app/src/main/java/com/example/recipe_app_compose/features/categories/presentation/view/FavoritesScreen.kt
@@ -1,0 +1,27 @@
+package com.example.recipe_app_compose.features.categories.presentation.view
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.Scaffold
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.sp
+import com.example.recipe_app_compose.core.components.VerticalScrollingWithFixedHeightDetail
+
+@Composable
+fun FavoritesScreen(modifier: Modifier) {
+    Scaffold(modifier = modifier.fillMaxSize()) { paddingValues ->
+        Column(
+            modifier = modifier
+                .padding(paddingValues)
+                .fillMaxSize(),
+            horizontalAlignment = Alignment.CenterHorizontally,
+            verticalArrangement = Arrangement.Center
+        ) {
+            VerticalScrollingWithFixedHeightDetail("Favorites Screen", 40.sp)
+        }
+    }
+}
