+++ app/src/main/java/com/example/recipe_app_compose/features/location/domain/model/LocationData.kt
@@ -1,0 +1,6 @@
+package com.example.recipe_app_compose.features.location.domain.model
+
+data class LocationData(
+    val latitude: Double,
+    val longitude: Double
+)
