--- app/src/main/java/com/example/recipe_app_compose/presentation/Widgets.kt
@@ -1,249 +1,0 @@
-package com.example.recipe_app_compose.presentation
-
-import androidx.compose.foundation.Image
-import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.PaddingValues
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.RowScope
-import androidx.compose.foundation.layout.WindowInsets
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.windowInsetsPadding
-import androidx.compose.foundation.rememberScrollState
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.foundation.verticalScroll
-import androidx.compose.material3.AlertDialog
-import androidx.compose.material3.BottomAppBarDefaults
-import androidx.compose.material3.ButtonDefaults
-import androidx.compose.material3.Card
-import androidx.compose.material3.CardDefaults
-import androidx.compose.material3.ElevatedButton
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Surface
-import androidx.compose.material3.Text
-import androidx.compose.material3.TextButton
-import androidx.compose.material3.contentColorFor
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.painter.Painter
-import androidx.compose.ui.layout.ContentScale
-import androidx.compose.ui.text.TextStyle
-import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.unit.Dp
-import androidx.compose.ui.unit.TextUnit
-import androidx.compose.ui.unit.dp
-import androidx.compose.ui.window.Dialog
-import androidx.compose.ui.window.DialogProperties
-import com.example.recipe_app_compose.presentation.view.IngredientScreen
-import com.example.recipe_app_compose.presentation.view.RandomMealPage
-
-@Composable
-fun FullScreenDialog(onDismissRequest: () -> Unit) {
-    Dialog(
-        onDismissRequest = { onDismissRequest() },
-        properties = DialogProperties(
-            usePlatformDefaultWidth = false,
-            dismissOnBackPress = true,
-        ),
-    ) {
-        Surface(
-            modifier = Modifier
-                .fillMaxSize()
-                .background(MaterialTheme.colorScheme.surfaceVariant),
-        ) {
-            RandomMealPage(modifier = Modifier.fillMaxSize())
-        }
-    }
-}
-
-@Composable
-fun IngredientFullScreenDialog(onDismissRequest: () -> Unit) {
-    Dialog(
-        onDismissRequest = { onDismissRequest() },
-        properties = DialogProperties(
-            usePlatformDefaultWidth = false,
-            dismissOnBackPress = true,
-        ),
-    ) {
-        Surface(
-            modifier = Modifier
-                .fillMaxWidth()
-                .height(600.dp)
-                .background(MaterialTheme.colorScheme.surfaceVariant),
-        ) {
-            IngredientScreen(modifier = Modifier.fillMaxSize())
-        }
-    }
-}
-
-@Composable
-fun AlertDialogExample(
-    onDismissRequest: () -> Unit,
-    onConfirmation: () -> Unit,
-    dialogTitle: String,
-    dialogText: String
-) {
-    AlertDialog(
-        title = {
-            Text(text = dialogTitle)
-        },
-        text = {
-            Text(text = dialogText)
-        },
-        onDismissRequest = {
-            onDismissRequest()
-        },
-        confirmButton = {
-            TextButton(
-                onClick = {
-                    onConfirmation()
-                }
-            ) {
-                Text("Confirm")
-            }
-        },
-        dismissButton = {
-            TextButton(
-                onClick = {
-                    onDismissRequest()
-                }
-            ) {
-                Text("Dismiss")
-            }
-        }
-    )
-}
-
-@Composable
-fun DialogWithImage(
-    onDismissRequest: () -> Unit,
-    onConfirmation: () -> Unit,
-    painter: Painter,
-    imageDescription: String,
-    text: String,
-    modifier: Modifier
-) {
-    Dialog(onDismissRequest = { onDismissRequest() }) {
-        // Draw a rectangle shape with rounded corners inside the dialog
-        Card(
-            modifier = modifier
-                .fillMaxWidth()
-                .height(450.dp)
-                .padding(16.dp),
-            shape = RoundedCornerShape(16.dp),
-            elevation = CardDefaults.cardElevation(10.dp)
-        ) {
-            Column(
-                modifier = modifier
-                    .fillMaxSize(),
-                verticalArrangement = Arrangement.Center,
-                horizontalAlignment = Alignment.CenterHorizontally,
-            ) {
-                Image(
-                    painter = painter,
-                    contentDescription = imageDescription,
-                    contentScale = ContentScale.Crop,
-                    modifier = modifier
-                        .height(300.dp)
-                )
-                Text(
-                    text = text, fontWeight = FontWeight.Medium,
-                    modifier = modifier.padding(10.dp),
-                )
-                Row(
-                    modifier = modifier
-                        .fillMaxWidth(),
-                    horizontalArrangement = Arrangement.Center,
-                ) {
-                    ElevatedButton(
-                        onClick = { onDismissRequest() },
-                        modifier = modifier.padding(5.dp),
-                        elevation = ButtonDefaults.buttonElevation(15.dp)
-                    ) {
-                        Text("Dismiss")
-                    }
-                    ElevatedButton(
-                        onClick = { onConfirmation() },
-                        modifier = modifier.padding(5.dp),
-                        elevation = ButtonDefaults.buttonElevation(15.dp)
-                    ) {
-                        Text("Confirm")
-                    }
-                }
-            }
-        }
-    }
-}
-
-@Composable
-fun MyBottomAppBar(
-    modifier: Modifier = Modifier,
-    containerColor: Color = BottomAppBarDefaults.containerColor,
-    contentColor: Color = contentColorFor(containerColor),
-    tonalElevation: Dp = BottomAppBarDefaults.ContainerElevation,
-    contentPadding: PaddingValues = BottomAppBarDefaults.ContentPadding,
-    windowInsets: WindowInsets = BottomAppBarDefaults.windowInsets,
-    content: @Composable RowScope.() -> Unit = {}
-) {
-    Surface(
-        color = containerColor,
-        contentColor = contentColor,
-        tonalElevation = tonalElevation,
-        shape = RoundedCornerShape(40.dp),
-        modifier = modifier
-    ) {
-        Row(
-            Modifier
-                .fillMaxWidth()
-                .windowInsetsPadding(windowInsets)
-                .height(70.0.dp)
-                .padding(contentPadding),
-            horizontalArrangement = Arrangement.SpaceEvenly,
-            verticalAlignment = Alignment.CenterVertically,
-            content = content
-        )
-    }
-}
-
-@Composable
-fun VerticalScrollingWithFixedHeightTextDemo(randomText: String) {
-    Text(
-        text = randomText,
-        style = TextStyle(fontWeight = FontWeight.Normal),
-        modifier = Modifier
-            .height(100.dp)
-            .verticalScroll(rememberScrollState())
-            .padding(top = 5.dp)
-    )
-}
-
-@Composable
-fun VerticalScrollingWithFixedHeightDetail(randomText: String, size: TextUnit) {
-    Text(
-        text = randomText,
-        style = TextStyle(fontWeight = FontWeight.Normal),
-        fontSize = size,
-        modifier = Modifier
-            .height(300.dp)
-            .verticalScroll(rememberScrollState())
-            .padding(top = 5.dp)
-    )
-}
-
-@Composable
-fun MessageCard(msg: String) {
-    Row(modifier = Modifier.padding(top = 2.dp, bottom = 2.dp)) {
-        Column {
-            Text(
-                text = msg,
-                style = TextStyle(fontWeight = FontWeight.Normal)
-            )
-        }
-    }
-}
