--- app/src/main/java/com/example/recipe_app_compose/presentation/view/InfoScreen.kt
@@ -1,27 +1,0 @@
-package com.example.recipe_app_compose.presentation.view
-
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.padding
-import androidx.compose.material3.Scaffold
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.unit.sp
-import com.example.recipe_app_compose.presentation.VerticalScrollingWithFixedHeightDetail
-
-@Composable
-fun InfoScreen(modifier: Modifier) {
-    Scaffold(modifier = modifier.fillMaxSize()) { paddingValues ->
-        Column(
-            modifier = modifier
-                .padding(paddingValues)
-                .fillMaxSize(),
-            horizontalAlignment = Alignment.CenterHorizontally,
-            verticalArrangement = Arrangement.Center
-        ) {
-            VerticalScrollingWithFixedHeightDetail("Info Screen", 40.sp)
-        }
-    }
-}
