--- app/src/main/java/com/example/recipe_app_compose/presentation/viewmodel/LocationViewModel.kt
@@ -1,15 +1,0 @@
-package com.example.recipe_app_compose.presentation.viewmodel
-
-import androidx.lifecycle.ViewModel
-import com.example.recipe_app_compose.domain.model.location.LocationData
-import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.StateFlow
-
-class LocationViewModel() : ViewModel() {
-    private val _location = MutableStateFlow<LocationData?>(null)
-    val location: StateFlow<LocationData?> = _location
-
-    val updateLocation: (LocationData) -> Unit = { newLocation ->
-        _location.value = newLocation
-    }
-}
